Replace EntityManager with EntityManagerInterface
Making a small change so that the worker type hints the interface rather than the class. This is to support alternative implementations of the EM. (In my case I'm using a proxy object. I should be able to just implement the interface, but I need to extend the EM in order for this worker to use it.)